# openroute/docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenRoute 面板 API
  version: 1.0.0
  description: |
    OpenRoute 是一个自研、可私有化的 TCP/UDP 转发管理面板。

    ## 认证
    支持四种方式：Session Cookie、JWT Bearer、API Token（`X-API-Key`）、
    节点 Token（`X-Node-Token`）。详见 `docs/API.md`。

    ## 统一响应体
    所有接口返回同一个外层结构，`code` 为 0 表示成功，
    非 0 时 `message` 为中文错误信息，`details` 携带字段级定位信息。

    ## 版本化
    管理接口全部位于 `/api/v1/`；节点通信接口位于 `/api/node/`（不带版本号）。
  license:
    name: 个人自用项目
servers:
  - url: http://127.0.0.1:18888
    description: 默认监听地址（可在 config.yml 中修改）
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
  - cookieAuth: []
paths:
...
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: 登录
      description: |
        校验用户名密码并签发访问令牌与刷新令牌。
        连续失败 3 次后要求图形验证码；10 次失败锁定 15 分钟。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  format: password
                captcha:
                  type: string
                  description: 图形验证码（连续失败 3 次后必填）
                captcha_id:
                  type: string
                  description: 验证码标识，取自 /auth/captcha
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiBody'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      summary: 刷新访问令牌
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiBody'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/me:
    get:
      tags: [auth]
      summary: 当前登录用户信息与权限
      responses:
        '200':
          description: 用户信息
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiBody'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/password:
    post:
      tags: [auth]
      summary: 修改自己的密码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [old_password, new_password]
              properties:
                old_password:
                  type: string
                  format: password
                new_password:
                  type: string
                  format: password
                  minLength: 6
      responses:
        '200':
          $ref: '#/components/responses/OK'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/nodes:
    get:
      tags: [node]
      summary: 节点列表
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Order'
        - name: group_id
          in: query
          schema: { type: integer, format: int64 }
        - name: online
          in: query
          schema: { type: boolean }
        - name: keyword
          in: query
          schema: { type: string }
      responses:
        '200':
          description: 节点列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiBody'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: '#/components/schemas/Node'
                          pagination:
                            $ref: '#/components/schemas/Pagination'
    post:
      tags: [node]
      summary: 创建节点
      description: 返回一次性 token 与可直接执行的安装命令（一键对接）。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeCreateInput'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiBody'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/NodeCreateResult'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/nodes/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
    get:
      tags: [node]
      summary: 节点详情
      responses:
        '200':
          description: 节点
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiBody'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Node'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [node]
      summary: 更新节点
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeUpdateInput'
      responses:
        '200':
          description: 更新后的节点
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiBody'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Node'
    delete:
      tags: [node]
      summary: 删除节点
      description: 删除后该节点上的规则会被标记为「节点已删除」并进入待处理清单。
      responses:
        '200':
          $ref: '#/components/responses/OK'
  /api/v1/nodes/{id}/install-command:
    post:
      tags: [node]
      summary: 生成安装命令
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      responses:
        '200':
          description: 安装命令与 token
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiBody'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          token: { type: string }
                          install_command: { type: string }
  /api/v1/nodes/{id}/upgrade:
    post:
      tags: [node]
      summary: 升级节点客户端
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      responses:
        '200':
          $ref: '#/components/responses/OK'
        '502':
          $ref: '#/components/responses/NodeError'
  /api/v1/nodes/{id}/restart:
    post:
      tags: [node]
      summary: 重启节点服务
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      responses:
        '200':
          $ref: '#/components/responses/OK'
        '502':
          $ref: '#/components/responses/NodeError'
  /api/v1/nodes/{id}/exec:
    post:
      tags: [node]
      summary: 在节点执行命令
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [command]
              properties:
                command: { type: string }
      responses:
        '200':
          description: 任务已下发
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiBody'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          task_id: { type: integer, format: int64 }
        '403':
          description: 节点设置了 DISABLE_EXECUTE=1，拒绝执行
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiBody'
  /api/v1/nodes/{id}/reset-token:
    post:
      tags: [node]
      summary: 重置节点密钥
      description: 重置后节点需要重新安装才能重新上线。
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      responses:
        '200':
          description: 新 token
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiBody'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          token: { type: string }
...
  /api/v1/nodes/{id}/rules:
    get:
      tags: [node]
      summary: 该节点上运行的规则
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      responses:
        '200':
          description: 规则列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiBody'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ForwardRule'
  /api/v1/nodes/{id}/logs:
    get:
      tags: [node]
      summary: 节点日志
      parameters:
        - $ref: '#/components/parameters/ResourceID'
        - name: lines
          in: query
          schema: { type: integer, default: 200, maximum: 2000 }
      responses:
        '200':
          description: 日志行
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiBody'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          lines:
                            type: array
                            items: { type: string }
  /api/v1/nodes/{id}/drift:
    get:
      tags: [node]
      summary: 配置漂移详情
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      responses:
        '200':
          description: 漂移详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiBody'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          drifted: { type: boolean }
                          details: {}
  /api/v1/nodes/{id}/drift/fix:
    post:
      tags: [node]
      summary: 一键纠正配置漂移
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      responses:
        '200':
          $ref: '#/components/responses/OK'
  /api/v1/nodes/batch/upgrade:
    post:
      tags: [node]
      summary: 批量升级节点
      requestBody:
        $ref: '#/components/requestBodies/BatchIDs'
      responses:
        '200':
          $ref: '#/components/responses/BatchResult'
  /api/v1/nodes/batch/exec:
    post:
      tags: [node]
      summary: 批量在节点执行命令
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids, command]
              properties:
                ids:
                  type: array
                  items: { type: integer, format: int64 }
                command: { type: string }
      responses:
        '200':
          $ref: '#/components/responses/BatchResult'
...
  /api/v1/node-groups/{id}/export:
    get:
      tags: [node-group]
      summary: 导出分组内节点为 CSV
      parameters:
        - $ref: '#/components/parameters/ResourceID'
      responses:
        '200':
          description: CSV 文件
          content:
            text/csv:
              schema:
                type: string

  # ───────────────────────── 设备组 ─────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 访问令牌，由 /auth/login 或 /auth/refresh 获取
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Token，创建时返回一次明文
    cookieAuth:
      type: apiKey
      in: cookie
      name: or_session
      description: WebUI 会话 Cookie（HttpOnly、SameSite=Lax）
    nodeTokenAuth:
      type: apiKey
      in: header
      name: X-Node-Token
      description: 节点密钥，由面板在创建节点时下发
...
  responses:
...
    Unauthorized:
      description: 未登录或令牌无效（40101 / 40102 / 40103 / 40104）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiBody'
...
  schemas:
    ApiBody:
      type: object
      properties:
        code:
          type: integer
          description: 0 表示成功，非 0 见错误码字典
        message:
          type: string
        data: {}
        details:
          $ref: '#/components/schemas/ErrorDetails'
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
...